fix: adds component for fileIcon and other fixes
