+++ package.json
@@ -30,8 +30,6 @@
     "@types/dompurify": "^3.0.5",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^22.10.1",
-    "@types/pdfjs-dist": "^2.10.377",
-    "@types/pretty-bytes": "^5.1.0",
     "@types/uuid": "^9.0.8",
     "bestzip": "^2.2.0",
     "concurrently": "^8.2.2",
@@ -47,9 +45,5 @@
     "tslib": "^2.6.2",
     "typescript": "^5.4.4"
   },
-  "dependencies": {
-    "@svelte-put/clickoutside": "^4.0.0",
-    "pdfjs-dist": "^4.9.155",
-    "pretty-bytes": "^6.1.1"
-  }
+  "dependencies": {}
 }
