+++ ui/package.json
@@ -28,6 +28,8 @@
     "linkify-string": "^4.2.0",
     "lodash-es": "^4.17.21",
     "p-retry": "^6.2.0",
+    "pdfjs-dist": "^4.9.155",
+    "pretty-bytes": "^6.1.1",
     "svelte-french-toast": "^1.2.0",
     "svelte-gestures": "^5.0.6",
     "svelte-persisted-store": "^0.12.0",
@@ -57,6 +59,7 @@
     "@types/dompurify": "^3.0.5",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^20.11.19",
+    "@types/pdfjs-dist": "^2.10.377",
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^8.16.0",
     "@typescript-eslint/parser": "^8.16.0",
